requirements.txt
pandas==1.5.3
Pillow==9.5.0
streamlit==1.21.0
torch==2.0.0+cu117
torchvision==0.15.1+cu117
